Made ByteChunkProvider.offset a long

+++ src/main/java/com/hierynomus/smbj/io/ByteChunkProvider.java
@@ -24,7 +24,7 @@ import java.io.OutputStream;
 public abstract class ByteChunkProvider {
     static final int CHUNK_SIZE = 64 * 1024;
 
-    private int offset;
+    private long offset;
 
     public abstract boolean isAvailable();
 
@@ -63,7 +63,7 @@ public abstract class ByteChunkProvider {
         }
     }
 
-    public int getOffset() {
+    public long getOffset() {
         return offset;
     }
